Module CCArrayLabels.Infix

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

a >>= f is the infix version of flat_map.

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

a >>| f is the infix version of map.

since

0.8

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

a >|= f is the infix version of map.

since

0.8

val (--) : int -> int -> int t

x -- y creates an array containing integers in the range x .. y. Bounds included.

val (--^) : int -> int -> int t

x --^ y creates an array containing integers in the range x .. y. Right bound excluded.

since

0.17

Let operators on OCaml >= 4.08.0, nothing otherwise

since

2.8

include CCShimsMkLet_.S with type 'a t_let := 'a array

type 'a t_let

val let+ : 'a t_let -> ('a -> 'b) -> 'b t_let

val and+ : 'a t_let -> 'b t_let -> ('a * 'b) t_let

val let* : 'a t_let -> ('a -> 'b t_let) -> 'b t_let

val and* : 'a t_let -> 'b t_let -> ('a * 'b) t_let