Module CCChar

Utils around char

Characters

type t = char

An alias for the type of characters.

val code : char -> int

Return the integer code of the argument.

val chr : int -> char

Return the character with the given integer code.

raises Invalid_argument
if the argument is outside the range [0x00;0xFF].

val escaped : char -> string

Return a string representing the given character, with special characters escaped following the lexical conventions of OCaml. All characters outside the ASCII printable range [0x20;0x7E] are escaped, as well as backslash, double-quote, and single-quote.

Predicates and comparisons

See also the Ascii module.

val compare : t -> t -> int

The comparison function for characters, with the same specification as Stdlib.compare. Along with the type t, this function compare allows the module Char to be passed as argument to the functors Set.Make and Map.Make.

val equal : t -> t -> bool

The equal function for chars.

ASCII characters

module Ascii = Stdlib.Char.Ascii

ASCII character set support.

val lowercase_ascii : char -> char

Use the equivalent Ascii.lowercase instead.

val uppercase_ascii : char -> char

Use the equivalent Ascii.uppercase instead.

Hashing

val seeded_hash : int -> t -> int

A seeded hash function for characters, with the same output value as Hashtbl.seeded_hash. This function allows this module to be passed as argument to the functor Hashtbl.MakeSeeded.

val hash : t -> int

An unseeded hash function for characters, with the same output value as Hashtbl.hash. This function allows this module to be passed as argument to the functor Hashtbl.Make.

val of_int_exn : int -> t

Alias to Char.chr. Return the character with the given ASCII code.

raises Invalid_argument
if the int is not within 0 … 255.

val of_int : int -> t option

Safe version of of_int_exn.

val to_int : t -> int

Alias to Char.code. Return the ASCII code of the argument.

val to_string : t -> string

to_string c returns a string containing c

val pp_buf : Stdlib.Buffer.t -> t -> unit

Renamed from pp since 2.0.

val pp : Stdlib.Format.formatter -> t -> unit

Renamed from print since 2.0.

val is_uppercase_ascii : t -> bool

is_uppercase_ascii c is true exactly when c is an uppercase ASCII character, i.e. '\064' < c < '\091'.

val is_lowercase_ascii : t -> bool

is_lowercase_ascii c is true exactly when c is a lowercase ASCII character, i.e. '\096' < c < '\123'.

val is_letter_ascii : t -> bool

is_letter_ascii c is true exactly when c is an ASCII letter, i.e. is_uppercase_ascii c || is_lowercase_ascii c.

val is_digit_ascii : t -> bool

is_digit_ascii c is true exactly when c is an ASCII digit, i.e. '\047' < c < '\058'.

val is_whitespace_ascii : t -> bool

is_whitespace_ascii c is true exactly when c is an ASCII whitespace character as defined by Unicode, i.e. either c = ' ' or '\008' < c < '\014'.

Infix Operators

module Infix : sig ... end

include module type of Infix

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (>) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool