Module CCFun.Infix

val (%>) : ('a -> 'b) -> ('b -> 'c) -> 'a -> 'c

(f %> g) x or (%>) f g x is g (f x). Alias to compose.

val (%) : ('b -> 'c) -> ('a -> 'b) -> 'a -> 'c

(f % g) x or (%) f g x is f (g x). Mathematical composition.

val (let@) : ('a -> 'b) -> 'a -> 'b

let@ x = foo in bar is the equivalent of foo @@ fun x -> bar. It can be very useful for resource management, alongside with protect.

val (||>) : ('a * 'b) -> ('a -> 'b -> 'c) -> 'c

x ||> f is f (fst x) (snd x)

val (|||>) : ('a * 'b * 'c) -> ('a -> 'b -> 'c -> 'd) -> 'd

like ||> but for tuples of size 3