Module CCSeq

Helpers for the standard Seq type

See oseq for a richer API.

type 'a iter = ('a -> unit) -> unit

type 'a gen = unit -> 'a option

type 'a equal = 'a -> 'a -> bool

type 'a ord = 'a -> 'a -> int

type 'a printer = Format.formatter -> 'a -> unit

Basics

type !'a t = unit -> 'a node

and !'a node =

| Nil

| Cons of 'a * 'a t

val is_empty : 'a t -> bool

val uncons : 'a t -> ('a * 'a t) option

val length : 'a t -> int

val iter : ('a -> unit) -> 'a t -> unit

val iteri : (int -> 'a -> unit) -> 'a t -> unit

val fold_lefti : ('acc -> int -> 'a -> 'acc) -> 'acc -> 'a t -> 'acc

val for_all : ('a -> bool) -> 'a t -> bool

val exists : ('a -> bool) -> 'a t -> bool

val find : ('a -> bool) -> 'a t -> 'a option

val find_index : ('a -> bool) -> 'a t -> int option

val find_map : ('a -> 'b option) -> 'a t -> 'b option

val find_mapi : (int -> 'a -> 'b option) -> 'a t -> 'b option

val iter2 : ('a -> 'b -> unit) -> 'a t -> 'b t -> unit

val fold_left2 : ('acc -> 'a -> 'b -> 'acc) -> 'acc -> 'a t -> 'b t -> 'acc

val for_all2 : ('a -> 'b -> bool) -> 'a t -> 'b t -> bool

val exists2 : ('a -> 'b -> bool) -> 'a t -> 'b t -> bool

val empty : 'a t

val return : 'a -> 'a t

val cons : 'a -> 'a t -> 'a t

val init : int -> (int -> 'a) -> 'a t

val unfold : ('b -> ('a * 'b) option) -> 'b -> 'a t

val forever : (unit -> 'a) -> 'a t

val cycle : 'a t -> 'a t

val iterate : ('a -> 'a) -> 'a -> 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

val mapi : (int -> 'a -> 'b) -> 'a t -> 'b t

val filter : ('a -> bool) -> 'a t -> 'a t

val filter_map : ('a -> 'b option) -> 'a t -> 'b t

val scan : ('b -> 'a -> 'b) -> 'b -> 'a t -> 'b t

val take : int -> 'a t -> 'a t

val drop : int -> 'a t -> 'a t

val take_while : ('a -> bool) -> 'a t -> 'a t

val drop_while : ('a -> bool) -> 'a t -> 'a t

val group : ('a -> 'a -> bool) -> 'a t -> 'a t t

exception Forced_twice

val once : 'a t -> 'a t

val transpose : 'a t t -> 'a t t

val append : 'a t -> 'a t -> 'a t

val concat : 'a t t -> 'a t

val flat_map : ('a -> 'b t) -> 'a t -> 'b t

val concat_map : ('a -> 'b t) -> 'a t -> 'b t

val zip : 'a t -> 'b t -> ('a * 'b) t

val map2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val interleave : 'a t -> 'a t -> 'a t

val sorted_merge : ('a -> 'a -> int) -> 'a t -> 'a t -> 'a t

val product : 'a t -> 'b t -> ('a * 'b) t

val map_product : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val unzip : ('a * 'b) t -> 'a t * 'b t

val split : ('a * 'b) t -> 'a t * 'b t

val partition_map : ('a -> ('b, 'c) Either.t) -> 'a t -> 'b t * 'c t

val partition : ('a -> bool) -> 'a t -> 'a t * 'a t

val of_dispenser : (unit -> 'a option) -> 'a t

val to_dispenser : 'a t -> unit -> 'a option

val ints : int -> int t

val nil : 'a t

val singleton : 'a -> 'a t

val repeat : ?n:int -> 'a -> 'a t

repeat ~n x repeats x n times then stops. If n is omitted, then x is repeated forever.

val head : 'a t -> 'a option

Head of the list.

val head_exn : 'a t -> 'a

Unsafe version of head.

raises Not_found
if the list is empty.

val tail : 'a t -> 'a t option

Tail of the list.

val tail_exn : 'a t -> 'a t

Unsafe version of tail.

raises Not_found
if the list is empty.

val equal : 'a equal -> 'a t equal

Equality step by step. Eager.

val compare : 'a ord -> 'a t ord

Lexicographic comparison. Eager.

val fold : ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

Fold on values.

val fold_left : ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

Alias for fold

val foldi : ('a -> int -> 'b -> 'a) -> 'a -> 'b t -> 'a

fold_lefti f init xs applies f acc i x where acc is the result of the previous computation or init for the first one, i is the index in the sequence (starts at 0) and x is the element of the sequence.

val filteri : (int -> 'a -> bool) -> 'a t -> 'a t

Similar to filter but the predicate takes aditionally the index of the elements.

val fmap : ('a -> 'b option) -> 'a t -> 'b t

Alias of filter_map.

val product_with : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Fair product of two (possibly infinite) lists into a new list. Lazy. The first parameter is used to combine each pair of elements.

val uniq : 'a equal -> 'a t -> 'a t

uniq eq l returns l but removes consecutive duplicates. Lazy. In other words, if several values that are equal follow one another, only the first of them is kept.

val flatten : 'a t t -> 'a t

Alias of concat

val range : int -> int -> int t

val (--) : int -> int -> int t

a -- b is the range of integers containing a and b (therefore, never empty).

val (--^) : int -> int -> int t

a --^ b is the integer range from a to b, where b is excluded.

Operations on two Collections

val fold2 : ('acc -> 'a -> 'b -> 'acc) -> 'acc -> 'a t -> 'b t -> 'acc

Alias for fold_left2.

val merge : 'a ord -> 'a t -> 'a t -> 'a t

Alias of sorted_merge.

val zip_i : 'a t -> (int * 'a) t

zip_i seq zips the index of each element with the element itself.

Misc

val sort : cmp:'a ord -> 'a t -> 'a t

Eager sort. Require the iterator to be finite. O(n ln(n)) time and space.

val sort_uniq : cmp:'a ord -> 'a t -> 'a t

Eager sort that removes duplicate values. Require the iterator to be finite. O(n ln(n)) time and space.

val memoize : 'a t -> 'a t

Avoid recomputations by caching intermediate results.

Fair Combinations

val fair_flat_map : ('a -> 'b t) -> 'a t -> 'b t

Fair version of flat_map.

val fair_app : ('a -> 'b) t -> 'a t -> 'b t

Fair version of (<*>).

Implementations

val pure : 'a -> 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

val (<*>) : ('a -> 'b) t -> 'a t -> 'b t

val (>>-) : 'a t -> ('a -> 'b t) -> 'b t

Infix version of fair_flat_map.

val (<.>) : ('a -> 'b) t -> 'a t -> 'b t

Infix version of fair_app.

Infix operators

module Infix : sig ... end

module type MONAD = sig ... end

module Traverse (M : MONAD) : sig ... end

Conversions

val of_list : 'a list -> 'a t

val to_list : 'a t -> 'a list

Gather all values into a list.

val of_array : 'a array -> 'a t

Iterate on the array.

val to_array : 'a t -> 'a array

Convert into array.

val to_rev_list : 'a t -> 'a list

Convert to a list, in reverse order. More efficient than to_list.

val to_iter : 'a t -> 'a iter

val to_gen : 'a t -> 'a gen

val of_gen : 'a gen -> 'a t

of_gen g consumes the generator and caches intermediate results.

val of_string : string -> char t

Iterate on characters.

IO

val pp : ?pp_start:unit printer -> ?pp_stop:unit printer -> ?pp_sep:unit printer -> 'a printer -> 'a t printer

pp ~pp_start ~pp_stop ~pp_sep pp_item ppf s formats the sequence s on ppf. Each element is formatted with pp_item, pp_start is called at the beginning, pp_stop is called at the end, pp_sep is called between each elements. By defaults pp_start and pp_stop does nothing and pp_sep defaults to (fun out -> Format.fprintf out ",@ ").