Module Containers_pp.Pretty

Pretty-printing.

These functions are parametrized by a width, and will try to fit the result within this width.

val to_out : width:int -> Out.t -> t -> unit

Render to an arbitrary output.

val to_string : width:int -> t -> string

Render to a string.

val to_buffer : width:int -> Stdlib.Buffer.t -> t -> unit

Render to a buffer.

val to_format : width:int -> Stdlib.Format.formatter -> t -> unit