Module CCArray.Infix

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

a >>= f is the infix version of flat_map.

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

a >>| f is the infix version of map.

val (>|=) : 'a t -> ('a -> 'b) -> 'b t

a >|= f is the infix version of map.

val (--) : int -> int -> int t

x -- y creates an array containing integers in the range x .. y. Bounds included.

val (--^) : int -> int -> int t

x --^ y creates an array containing integers in the range x .. y. Right bound excluded.

Let operators on OCaml >= 4.08.0, nothing otherwise

val let+ : 'a array -> ('a -> 'b) -> 'b array

val and+ : 'a array -> 'b array -> ('a * 'b) array

val let* : 'a array -> ('a -> 'b array) -> 'b array

val and* : 'a array -> 'b array -> ('a * 'b) array