Module CCChar

Utils around char

val code : char -> int

Return the ASCII code of the argument.

val chr : int -> char

Return the character with the given ASCII code.

raises Invalid_argument
if the argument is outside the range 0--255.

val escaped : char -> string

Return a string representing the given character, with special characters escaped following the lexical conventions of OCaml. All characters outside the ASCII printable range (32..126) are escaped, as well as backslash, double-quote, and single-quote.

val lowercase : char -> char

Convert the given character to its equivalent lowercase character, using the ISO Latin-1 (8859-1) character set.

deprecated
Functions operating on Latin-1 character set are deprecated.

val uppercase : char -> char

Convert the given character to its equivalent uppercase character, using the ISO Latin-1 (8859-1) character set.

deprecated
Functions operating on Latin-1 character set are deprecated.

val lowercase_ascii : char -> char

Convert the given character to its equivalent lowercase character, using the US-ASCII character set.

val uppercase_ascii : char -> char

Convert the given character to its equivalent uppercase character, using the US-ASCII character set.

type t = char

An alias for the type of characters.

val equal : t -> t -> bool

The equal function for chars.

val compare : t -> t -> int

The comparison function for characters, with the same specification as Stdlib.compare. Along with the type t, this function compare allows the module Char to be passed as argument to the functors Set.Make and Map.Make.

val of_int_exn : int -> t

Alias to Char.chr. Return the character with the given ASCII code.

raises Invalid_argument
if the int is not within 0 … 255.

val of_int : int -> t option

Safe version of of_int_exn.

val to_int : t -> int

Alias to Char.code. Return the ASCII code of the argument.

val to_string : t -> string

to_string c returns a string containing c

val pp_buf : Stdlib.Buffer.t -> t -> unit

Renamed from pp since 2.0.

val pp : Stdlib.Format.formatter -> t -> unit

Renamed from print since 2.0.

Infix Operators

module Infix : sig ... end

include module type of Infix

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (>) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool