Module CCChar

Utils around char

since

0.14

Documentation for the standard Char module

include module type of sig ... end

val code : char -> int

val chr : int -> char

val escaped : char -> string

val lowercase : char -> char

val uppercase : char -> char

val lowercase_ascii : char -> char

val uppercase_ascii : char -> char

type t = char

val compare : t -> t -> int

val equal : t -> t -> bool

val unsafe_chr : int -> char

val compare : t -> t -> int

The comparison function for characters, with the same specification as Stdlib.compare. Along with the type t, this function compare allows the module Char to be passed as argument to the functors Set.Make and Map.Make.

val of_int_exn : int -> t

Alias to Char.chr. Return the character with the given ASCII code.

raises Invalid_argument

if the int is not within 0 … 255.

since

1.0

val of_int : int -> t option

Safe version of of_int_exn.

since

1.0

val to_int : t -> int

Alias to Char.code. Return the ASCII code of the argument.

since

1.0

val to_string : t -> string

to_string c returns a string containing c

since

2.7

val pp_buf : Stdlib.Buffer.t -> t -> unit

Renamed from pp since 2.0.

val pp : Stdlib.Format.formatter -> t -> unit

Renamed from print since 2.0.