Module CCResult.Infix

val (<$>) : ('a -> 'b) -> ('a, 'err) t -> ('b, 'err) t

Infix version of map.

since

3.0

val (>|=) : ('a, 'err) t -> ('a -> 'b) -> ('b, 'err) t

Infix version of map with reversed arguments.

val (>>=) : ('a, 'err) t -> ('a -> ('b, 'err) t) -> ('b, 'err) t

Monadic composition. e >>= f proceeds as f x if e is Ok x or returns e if e is an Error.

val (<*>) : ('a -> 'b, 'err) t -> ('a, 'err) t -> ('b, 'err) t

a <*> b evaluates a and b, and, in case of success, returns Ok (a b). Otherwise, it fails, and the error of a is chosen over the error of b if both fail.

Let operators on OCaml >= 4.08.0, nothing otherwise

since

2.8

include CCShimsMkLet_.S2 with type ('a, 'e) t_let2 := ('a, 'e) result

type ('a, 'e) t_let2

val let+ : ('a, 'e) t_let2 -> ('a -> 'b) -> ('b, 'e) t_let2

val and+ : ('a, 'e) t_let2 -> ('b, 'e) t_let2 -> ('a * 'b, 'e) t_let2

val let* : ('a, 'e) t_let2 -> ('a -> ('b, 'e) t_let2) -> ('b, 'e) t_let2

val and* : ('a, 'e) t_let2 -> ('b, 'e) t_let2 -> ('a * 'b, 'e) t_let2